# public/openApi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Places API",
        "description": "This is the full documentation for the Places API.",
        "contact": {
            "email": "nizarberjawi12@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Places API host"
        }
    ],
    "paths": {
...
        "/countries/{countryCode}/currency": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Returns the currency of a specific country",
                "description": "Display the currency of a country.",
                "operationId": "327a85bf68b6a59cd02e38af09185d25",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/countryCode"
                    },
                    {
                        "$ref": "#/components/parameters/currencyFilter"
                    },
                    {
                        "$ref": "#/components/parameters/currencyInclude"
                    },
                    {
                        "$ref": "#/components/parameters/currencySort"
                    },
                    {
                        "$ref": "#/components/parameters/pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/currency"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Country not found"
                    }
                }
            }
        },
...
        "/countries/{countryCode}/places": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Returns the places available in a specific country",
                "description": "Display the places available in a country.",
                "operationId": "a77e032151128f3b2c7f6fe22d030420",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/countryCode"
                    },
                    {
                        "$ref": "#/components/parameters/placeFilter"
                    },
                    {
                        "$ref": "#/components/parameters/placeInclude"
                    },
                    {
                        "$ref": "#/components/parameters/placeSort"
                    },
                    {
                        "$ref": "#/components/parameters/pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/place"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Country not found"
                    }
                }
            }
        },
...
        "/currencies": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Returns a list of paginated currencies",
                "description": "Display a listing of all currencies.",
                "operationId": "cc96e72a6ab0802a84471e6bd12939f3",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/currencyFilter"
                    },
                    {
                        "$ref": "#/components/parameters/currencyInclude"
                    },
                    {
                        "$ref": "#/components/parameters/currencySort"
                    },
                    {
                        "$ref": "#/components/parameters/pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/currency"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/currencies/{currencyCode}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Display a specified currency.",
                "operationId": "getCurrencyByCode",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/currencyCode"
                    },
                    {
                        "$ref": "#/components/parameters/currencyInclude"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/currency"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Currency not found"
                    }
                }
            }
        },
        "/featureClasses": {
            "get": {
                "tags": [
                    "Feature Classes"
                ],
                "summary": "Returns a list of paginated feature classes",
                "description": "Display a listing of all feature classes.",
                "operationId": "aed912c79cd30d506bcf9166c684072e",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/featureClassFilter"
                    },
                    {
                        "$ref": "#/components/parameters/featureClassInclude"
                    },
                    {
                        "$ref": "#/components/parameters/featureClassSort"
                    },
                    {
                        "$ref": "#/components/parameters/pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/featureClass"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/featureClasses/{featureClassCode}": {
            "get": {
                "tags": [
                    "Feature Classes"
                ],
                "summary": "Display a specified feature class.",
                "operationId": "getFeatureClassByCode",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/featureClassCode"
                    },
                    {
                        "$ref": "#/components/parameters/featureClassInclude"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/featureClass"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Feature class not found"
                    }
                }
            }
        },
...
        "/places/{geonameId}/alternateNames": {
            "get": {
                "tags": [
                    "Places"
                ],
                "summary": "Display the alternate names of a place.",
                "operationId": "getAlternateNamesForPlace",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/geonameId"
                    },
                    {
                        "$ref": "#/components/parameters/alternateNameFilter"
                    },
                    {
                        "$ref": "#/components/parameters/alternateNameInclude"
                    },
                    {
                        "$ref": "#/components/parameters/alternateNameSort"
                    },
                    {
                        "$ref": "#/components/parameters/pagination"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/location"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alternate name not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "currency": {
                "title": "Currency",
                "properties": {
                    "name": {
                        "description": "The name of the currency",
                        "type": "string",
                        "example": "Dollar"
                    },
                    "code": {
                        "description": "The code of the currency",
                        "type": "string",
                        "example": "AUD"
                    }
                },
                "type": "object"
            },
            "featureClass": {
                "title": "Feature Class",
                "properties": {
                    "code": {
                        "description": "The code of the feature class",
                        "type": "string",
                        "example": "A"
                    },
                    "description": {
                        "description": "The description of the feature class",
                        "type": "string",
                        "example": "Country, state, region,..."
                    }
                },
                "type": "object"
            },
...
            "place": {
                "title": "Place",
                "properties": {
                    "geonameId": {
                        "description": "The Geoname ID of a certain place",
                        "type": "integer",
                        "example": "105480"
                    },
                    "name": {
                        "description": "The name of a place",
                        "type": "string",
                        "example": "Jāl Jahlān"
                    },
                    "asciiName": {
                        "description": "The ASCII name of a place",
                        "type": "string",
                        "example": "Jal Jahlan"
                    },
                    "population": {
                        "description": "The population of a certain place",
                        "type": "integer",
                        "example": "0"
                    },
                    "elevation": {
                        "description": "The elevation of a place in meters",
                        "type": "integer",
                        "example": "0"
                    },
                    "dem": {
                        "description": "The digital elevation model of a place in meters",
                        "type": "integer",
                        "example": "79"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "countryCode": {
                "name": "countryCode",
                "in": "path",
                "description": "The iso3166Alpha2 code of the country",
                "required": true,
                "schema": {
                    "type": "string"
                },
                "example": "AU"
            },
            "currencyCode": {
                "name": "currencyCode",
                "in": "path",
                "description": "The ISO 4217 code of the currency",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "featureClassCode": {
                "name": "featureClassCode",
                "in": "path",
                "description": "The code of the Feature Class",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
            "geonameId": {
                "name": "geonameId",
                "in": "path",
                "description": "The Geoname ID of the place",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
            "alternateNameFilter": {
                "name": "filter",
                "in": "query",
                "description": "Filter alternate names by certain criteria",
                "required": false,
                "style": "deepObject",
                "schema": {
                    "properties": {
                        "isPreferredName": {
                            "properties": {
                                "eq": {
                                    "description": "The attributes we can use to filter.",
                                    "type": "boolean",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "type": "object",
                    "enum": [
                        "name",
                        "isPreferredName",
                        "isShortName",
                        "isHistoric",
                        "isColloquial",
                        "languageCode"
                    ]
                }
            },
            "alternateNameInclude": {
                "name": "include",
                "in": "query",
                "description": "Include resources related to the specified alternate name.",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "language",
                            "place"
                        ]
                    }
                }
            },
            "alternateNameSort": {
                "name": "sort",
                "in": "query",
                "description": "Sort the result set by certain properties.",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "name",
                            "isPreferredName",
                            "isShortName",
                            "isHistoric",
                            "isColloquial",
                            "-name",
                            "-isPreferredName",
                            "-isShortName",
                            "-isHistoric",
                            "-isColloquial"
                        ]
                    }
                }
            },
...
            "currencyFilter": {
                "name": "filter",
                "in": "query",
                "description": "Filter currencies by certain criteria",
                "required": false,
                "style": "deepObject",
                "schema": {
                    "properties": {
                        "code": {
                            "description": "The attributes we can use to filter.",
                            "type": "string",
                            "example": "AUD"
                        }
                    },
                    "type": "object",
                    "enum": [
                        "code",
                        "name"
                    ]
                }
            },
            "currencyInclude": {
                "name": "include",
                "in": "query",
                "description": "Include related resources",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "countries"
                        ]
                    }
                }
            },
            "currencySort": {
                "name": "sort",
                "in": "query",
                "description": "Sort the result set by certain properties.",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "code",
                            "name",
                            "-code",
                            "-name"
                        ]
                    }
                }
            },
            "featureClassFilter": {
                "name": "filter",
                "in": "query",
                "description": "Filter feature classes by certain criteria",
                "required": false,
                "style": "deepObject",
                "schema": {
                    "properties": {
                        "code": {
                            "description": "The attributes we can use to filter.",
                            "type": "string",
                            "example": "A"
                        }
                    },
                    "type": "object",
                    "enum": [
                        "code"
                    ]
                }
            },
            "featureClassInclude": {
                "name": "include",
                "in": "query",
                "description": "Include related resources",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "featureCodes"
                        ]
                    }
                }
            },
            "featureClassSort": {
                "name": "sort",
                "in": "query",
                "description": "Sort the result set by certain properties.",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "code",
                            "-code"
                        ]
                    }
                }
            },
...
            "placeFilter": {
                "name": "filter",
                "in": "query",
                "description": "Filter places by certain criteria",
                "required": false,
                "style": "deepObject",
                "schema": {
                    "properties": {
                        "populationGt": {
                            "description": "The attributes we can use to filter.",
                            "type": "integer",
                            "example": "100000"
                        }
                    },
                    "type": "object",
                    "enum": [
                        "name",
                        "featureCode",
                        "countryCode",
                        "elevation",
                        "elevationGt",
                        "elevationGte",
                        "elevationLt",
                        "elevationLte",
                        "population",
                        "populationGt",
                        "populationGte",
                        "populationLt",
                        "populationLte"
                    ]
                }
            },
            "placeInclude": {
                "name": "include",
                "in": "query",
                "description": "Include related resources",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "country",
                            "location",
                            "alternateNames",
                            "featureClass",
                            "featureCode",
                            "timeZone"
                        ]
                    }
                }
            },
            "placeSort": {
                "name": "sort",
                "in": "query",
                "description": "Sort the result set by certain properties.",
                "required": false,
                "explode": false,
                "schema": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "name",
                            "featureCode",
                            "countryCode",
                            "timeZoneCode",
                            "population",
                            "elevation",
                            "-name",
                            "-featureCode",
                            "-countryCode",
                            "-timeZoneCode",
                            "-population",
                            "-elevation"
                        ]
                    }
                }
            },
            "pagination": {
                "name": "page",
                "in": "query",
                "description": "Paginate the data",
                "required": false,
                "style": "deepObject",
                "schema": {
                    "properties": {
                        "cursor": {
                            "description": "The paginator used to paginate the result.",
                            "type": "string",
                            "example": "eyJpc28zMTY2X2FscGhhMiI6IkJPIiwiX3BvaW50c1RvTmV4dEl0ZW1zIjp0cnVlfQ"
                        },
                        "size": {
                            "type": "integer",
                            "example": "10"
                        }
                    },
                    "type": "object",
                    "enum": [
                        "number",
                        "size"
                    ]
                }
            },
...
        }
    },
...
}